Make KDD IPAM controller reactive to node deletions
