+++ glide.lock
@@ -1,5 +1,5 @@
-hash: 2cf285f0446f4aa3328efbe6133542c632b306f70a6a9261dd8b88f7aecae268
-updated: 2019-02-11T15:39:13.006712395-08:00
+hash: 93bf41b244a92b50b94d6044e3d9fc52f2d4822dfb7553700474ce14e16f8834
+updated: 2019-02-13T13:15:22.218401776-08:00
 imports:
 - name: cloud.google.com/go
   version: 3b1ae45394a234c385be014e9a488f2bb6eef821
